requirements.txt
numpy>=1.26
pandas>=2.2
scikit-learn>=1.4
PyYAML>=6.0
pytest>=8.2
pytest-cov>=5.0
matplotlib>=3.8
seaborn>=0.13
tqdm>=4.66
jupyter>=1.0
tensorflow>=2.15
xlrd>=2.0.1
